\name{plot.counts-package}
\alias{plot.counts-package}
\alias{plot.counts}
\docType{package}
\title{
plot.counts
}
\description{
Utility functions for comparing and elementary plotting of count data under different conditions
}
\details{
\tabular{ll}{
Package: \tab plot.counts\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-22\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{
Thomas Braschler

Maintainer: thomas.braschler@gmail.com
}
\keyword{ package }

